server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.124.4",
    "itsdangerous>=2.2.0",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "spotipy>=2.25.2",
    "uvicorn>=0.38.0",
]
